transcription/requirements.txt
openai-whisper
requests
watchdog